Make macOS executables universal, for both x86_64 and arm64
This provides a rebuild of all macOS executables. For reproducibility,
this was done entirely by `build_mac.sh`. I verified that different
machines produce bit-for-bit identical output. A build machine running
Xcode 16.2 16C5032a on macOS 15.2 24C101 was used.

These tools should run on macOS 10.13 and later. This version was
selected as it's the minimum macOS version that the xPack tools declare
support for.

A build of libusb is included, so that the provided `dfu-util` no longer
has an external dependency.

The universal structure was chosen for these tools to avoid changing
paths, anywhere they may be relied on. If desired, `build_mac.sh` can
also be used to produce single-architecture tools, even cross-compiling,
depending on the value of the `ARCH` variable.

The current released versions were selected:
 - libusb 1.0.27
 - dfu-util 0.11 (including dfu-prefix and dfu-suffix)
 - hid-flash from STM32_HID_Bootloader 2.2.2 (this program erroneously
   reports its version as 2.2.1), with a patch from
https://github.com/Serasidis/STM32_HID_Bootloader/issues/68#issuecomment-2009105851.
 - upload_reset from this repository at the HEAD of main

Link: https://github.com/stm32duino/BoardManagerFiles/issues/72
Signed-off-by: Mark Mentovai <mark@mentovai.com>